Move the environment wire shape into its own module
Review follow-up to a5e187882. Sharing the environment mapper by exporting it
from environments.ts meant deploymentPipelines.ts imported that file, which
imports appendEnvironmentToDefaultPipeline back — an import cycle. It worked,
because both sides are only referenced inside function bodies, but it left the
two largest modules in this directory mutually dependent and one refactor away
from a module-scope reference and a confusing TDZ failure.

Put BffEnvironment and toEnvironment in _environmentShape.ts instead. Both
domain modules depend on the shape module and neither depends on the other, so
the edge from environments.ts to deploymentPipelines.ts is one-directional
again. A sweep of all 57 modules under src/api/cloud now reports no cycles,
where the previous commit had the two-node cycle in both directions.

No behaviour change: the mapper is byte-identical and environments.ts simply
consumes it rather than exporting it, so nothing new reaches the public surface
via index.ts.
